+++ src/main/java/com/example/demo/controller/PregledController.java
@@ -20,10 +20,12 @@ import com.example.demo.model.AdministratorKlinike;
 import com.example.demo.model.Doktor;
 import com.example.demo.model.Klinika;
 import com.example.demo.model.Korisnik;
+import com.example.demo.model.LogedUser;
 import com.example.demo.model.Pregled;
 import com.example.demo.model.Sala;
 import com.example.demo.model.StatusPregleda;
 import com.example.demo.model.Termin;
+import com.example.demo.model.UlogaKorisnika;
 import com.example.demo.service.AdministratorKlinikeService;
 import com.example.demo.service.DoktorService;
 import com.example.demo.service.EmailService;
@@ -78,7 +80,13 @@ public class PregledController {
 	
 	@RequestMapping(value = "/izmeni", method = RequestMethod.PUT)
 	public ResponseEntity<PregledDTO> izmeni(@RequestBody PregledDTO pregledDTO) {
+		if(!LogedUser.getInstance().getUserRole().equals(UlogaKorisnika.LEKAR)) {
+			return new ResponseEntity<>(HttpStatus.BAD_REQUEST);
+		}
 		Pregled pregled = pregledService.findOne(pregledDTO.getId());
+		if (!pregled.getDoktor().getIdKorisnik().equals(LogedUser.getInstance().getUserId())) {
+			return new ResponseEntity<PregledDTO>(HttpStatus.BAD_REQUEST);
+		}
 		pregled.setNaziv(pregledDTO.getNaziv());
 		pregled.setAnamneza(pregledDTO.getAnamneza());
 		pregled.setTipPregleda(pregledDTO.getTipPregleda());
@@ -89,7 +97,13 @@ public class PregledController {
 	
 	@RequestMapping(value = "/obrisi/{id}", method = RequestMethod.DELETE)
 	public ResponseEntity<String> obrisi(@PathVariable("id") Long identifikacija) {
+		if(!LogedUser.getInstance().getUserRole().equals(UlogaKorisnika.LEKAR)) {
+			return new ResponseEntity<>(HttpStatus.BAD_REQUEST);
+		}
 		Pregled pregled = pregledService.findOne(identifikacija);
+		if (!pregled.getDoktor().getIdKorisnik().equals(LogedUser.getInstance().getUserId())) {
+			return new ResponseEntity<String>(HttpStatus.BAD_REQUEST);
+		}
 		pregledService.delete(pregled);
 		return new ResponseEntity<String>("", HttpStatus.OK);
 	}
